Adding some features
Added the Sum-Method, it calculates the sum of all values in the tree.
Also added the NumOfNodes Method, this method counts the number of nodes in the tree
Last but not least, these to methods are used in Average, this claculates the average value of the nodes in the tree.

<!-- Warning! -->
Because of time-reasons, these methds are not tested